requirements.txt
torch
transformers
librosa
pyannote.audio
sumy
openai-whisper
